+++ nncf/experimental/torch2/function_hook/weak_map.py
