+++ tests/torch2/function_hook/test_hook_storage.py
@@ -15,9 +15,9 @@ import pytest
 import torch
 from torch import nn
 
-from nncf.experimental.torch2.hook_executor.hook_storage import HookStorage
-from nncf.experimental.torch2.hook_executor.hook_storage import HookType
-from tests.torch2.hook_executor.helpers import CallCount
+from nncf.experimental.torch2.function_hook.hook_storage import HookStorage
+from nncf.experimental.torch2.function_hook.hook_storage import HookType
+from tests.torch2.function_hook.helpers import CallCount
 
 
 class CheckPriority(nn.Module):
