+++ tests/torch2/function_hook/test_wrapper.py
@@ -16,12 +16,12 @@ import onnxruntime as ort
 import pytest
 import torch
 
-from nncf.experimental.torch2.hook_executor.hook_storage import HookType
-from nncf.experimental.torch2.hook_executor.wrapper import is_wrapped
-from nncf.experimental.torch2.hook_executor.wrapper import register_post_function_hook
-from nncf.experimental.torch2.hook_executor.wrapper import register_pre_function_hook
-from nncf.experimental.torch2.hook_executor.wrapper import wrap_model
-from tests.torch2.hook_executor import helpers
+from nncf.experimental.torch2.function_hook.hook_storage import HookType
+from nncf.experimental.torch2.function_hook.wrapper import is_wrapped
+from nncf.experimental.torch2.function_hook.wrapper import register_post_function_hook
+from nncf.experimental.torch2.function_hook.wrapper import register_pre_function_hook
+from nncf.experimental.torch2.function_hook.wrapper import wrap_model
+from tests.torch2.function_hook import helpers
 
 ADD_VALUE = 2.0
